move file blueprint from yafowil.common to yafowil.file
